Centralise date handling and migrate image filenames to ISO format
  Introduce APODDate as source of truth for date logic:
  NASA-Eastern "today", invariant-culture yyyy-MM-dd formatting/parsing,
  and the 1995-06-16 inception date. Route APODCache, APODInfo,
  PictureData, and ViewModel through it, replacing ad-hoc
  DateTime.UtcNow and culture-sensitive ParseExact calls.

  This fixes two long-standing correctness bugs: APOD release dates were
  compared against UTC rather than NASA's Eastern timezone, and date
  (de)serialisation was not culture-invariant.

  Add FileMigration to rename legacy image/json files to the ISO naming
  scheme on startup (invoked from both the console app and the GUI).

  Also:
  - Add Directory.Build.props for date-based versioning
    (yyyy.MM.dd.{Patch}); null-coalesce version strings to "Unknown".
  - Add ConfigurableTimeoutHandler so per-request network timeouts honour
    config.NetworkTimeout.
  - Expose IConfigurationService via ViewModel.Config; split GetImagesSize
    into image-size and cache-size helpers with existence guards.

+++ APODWallpaper/Interfaces/IConfigurationService.cs
@@ -13,7 +13,7 @@ namespace APODWallpaper.Interfaces
         bool ExplainImage { get; set; }
         string BaseUrl { get; set; }
         string ConfiguratorTheme { get; set; }
-        int NetworkTimeout { get; set; }
+        long NetworkTimeout { get; set; }
         long PreviewQuality { get; set; }
         long WallpaperStyle { get; set; }
         string API_KEY { get; set; }
